experiments/repeatable/pyproject.toml
[tool.poetry]
name = "experiment"
version = "0.1.0"
description = ""
authors = ["Luke Channings <luke@channings.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
paho-mqtt = "^1.6.1"
python-dotenv = "^1.0.0"
docopt = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
